Use webpack 2 and move config to separate file
